+++ assets/js/steps/CheckInItems.jsx
@@ -49,6 +49,7 @@ function CheckInItems({ actionHandler }) {
       barcodeTimeout,
       soundEnabled,
       reservedMaterialInstruction,
+      otherPermanentLocationInstruction,
       hasPrinter,
       debugEnabled,
     },
@@ -190,19 +191,22 @@ function CheckInItems({ actionHandler }) {
    */
   useEffect(() => {
     if (soundEnabled && soundToPlay !== "") {
-      new Sound().playSound(soundToPlay);
+      let sound = new Sound();
+      sound.playSound(soundToPlay);
     }
   }, [soundEnabled, soundToPlay]);
 
   useEffect(() => {
     if (items && items.length > 0) {
       setDisplayedItems(
-        adaptListOfBooksToBanner(items, reservedMaterialInstruction).sort(
-          (a, b) => (a.timestamp < b.timestamp ? 1 : -1)
-        )
+        adaptListOfBooksToBanner(
+          items,
+          reservedMaterialInstruction,
+          otherPermanentLocationInstruction
+        ).sort((a, b) => (a.timestamp < b.timestamp ? 1 : -1))
       );
     }
-  }, [items, reservedMaterialInstruction]);
+  }, [items, otherPermanentLocationInstruction, reservedMaterialInstruction]);
 
   return (
     <>